Clear Email Address Field Before Filling in
During the Capybara end-to-end tests we enter the submission email
address. When running these tests from my machine Chrome was autofilling
the submission email box with the email address, then Capybara would
enter it again resulting in a concatenated and invalid email address.
This change ensures the box is first cleared before the email address is
entered.