Fix signing in with GOV.UK Signon for end to end tests
In the Sigon integration environment, the content for the page
requesting a TOTP code has changed; the field used to be labelled "Your
verification code", it's now labelled "Code from app". This broke our
end to end tests on dev, as we couldn't enter the 2nd factor auth token.

This commit fixes things by changing the line to look for the input
`name` text rather than label.