Fix end to end tests for PR alphagov/forms-admin#536
In alphagov/forms-admin#536 the flash message for deleting a form is
changed to a success banner. This means that the content is no longer
contained in a `p` but in a `div.govuk-notification-banner`.
Additionally, quotes are added around the form name.

This commit updates the spec to pass with these changes. The spec will
no longer pass on older versions of the app; however I think this is
fine, as they are normally only run during deployment.