Move helper methods out of spec file and into support file
I was finding it too hard to focus on the actual spec code with
everything going on in the file `spec/features/complete_spec.rb`.

This commit moves everything that isn't an example or example group into
a new helper file, `spec/support/feature_helpers.rb`. We could probably
split this file up a bit in future, but this change already feels like
an improvement to me.