Add more log messages
To make it easier to see where the tests have got up to at a high level,
this commit adds some extra logging (visible when LOG_LEVEL is set to
INFO or lower). I've tried to make the messages descriptive of the
features we're trying to test, I've used Gherkin syntax [[1]] to that end.

[1]: https://cucumber.io/docs/gherkin/reference/