Extract core dependency versions into build args
This commit adds build arguments [[1]] which define the base image to
use for our end to end test image.

These build arguments are not intended to be set by `docker build`
command line arguments, instead we just want to use the default values
for interpolation later in the Dockerfile. This makes it easier for the
`update_app_versions script` to update the versions of Alpine Linux and
Ruby in the Dockerfile (see alphagov/forms-deploy#503).

[1]: https://docs.docker.com/engine/reference/builder/#arg