Remove GOV.UK Notify retry logic
The logic for retrying email receipt seems to be broken, and appears to
have broken since it was originally committed in 4d74800. I think that
we never retry the call to `NotifyService#get_confirmation_code_email`,
so if no emails were returned this code would just wait for 5 seconds
before returning false.

The fact that we haven't seen a lot of failures from our end to end
tests due to failure to get emails from Notify suggests that we don't
actually need to do retries, and just waiting for 3 seconds is
sufficient.

So this commit removes the unneeded code, since it isn't being used and
isn't working. If we find we do need it in future we can always add
something else back in.