Rename spec files
This commit tries to make it easier to see which spec files relate to
end to end tests and which relate to smoke tests.

We put the different tests in different folders, so in future we can
have lots of different files related to each type of test without having
to worry about changing scripts or commands in the deploy repo.