Make e2e tests pass with group feature
The end to end tests will fail if the groups feature is enabled.

The groups feature changes the start of the journey.

To fix the tests, they must visit an existing, active group before
creating and deleting a form.

This commit adds a new method to visit the group, called "End to end
tests" and continues the journey from there.

A few assertions about the URL have been removed as they are no longer
true and are not needed.