Change form creation step for repeating pages
When the feature flag for repeating pages is enabled the end to end test
will fail when creating new pages.

This is because the feature flag enables a new radio input for allowing
multiple answers to a question. This input must be completed for the
page to be saved.

Currently the tests fail when the hit this point, so would fail in
environments where the flag is on.

This PR adds checks for the input when creating new pages for selection
and single line of text questions. When the input is present, it chooses
no. This ensures that the form filler part doesn't need to change.

The test is:

```ruby
page.has_field?("pages_question_input[is_repeatable]", type: :radio, visible: :all)
```

We need to use `visible: :all` or `false` because labels for radio
buttons in the govuk design system hide the original inputs, so for
capybara they are not visible.

The check is in two places and could be extracted to a method.

It's been inlined because it's only two places and it keeps the context
clear.