Remove conditional added for share preview
We added a conditional to the test so that it would work whether or not
the share preview task was present. We've now merged in the share
preview work, so it will appear on all forms. This means we can safely
remove the conditional since the task will always be present and we want
the test to fail if this isn't the case.