Fix tests for new route page
This commit to the admin changes the user joourney for adding routes to
a question.

This breaks the end to end tests becaues it doesn't expect this page to
be there.

This commit adds a check to the end to end tests to ensure if the page
is shown the user is returned to page list so the tests can continue as
normal.

This is conditional so that the tests work both with the new page and
without so that the change can be deployed.