Remove logic for handling old selection flow
In the helper for adding a selection question, we previously added a
condition for checking which selection journey was in use and proceeding
handling the different question flows.

We've now removed the old selection journey from forms-admin, so we can
remove the check and the code for the old journey.