Change Docker image to use older version of Chromium
We are seeing intermittent failures when end to end tests are run
locally. These failures are related to Selenium and chromedriver 134 [[1]].
They are false positives and do not mean the code is broken.

We want to make sure we don't have these issues with our end to end
tests and smoke tests in AWS. So far this hasn't happened, because our
Docker image hasn't been rebuilt recently, but the next time the image
is rebuilt we will start seeing failures, as the Alpine 3.21 release
branch has recently been updated with the latest version of Chromium.

This commit makes sure we won't get the affected version of
Chromium/chromedriver in our end to end test Docker image by rolling
back the base image to the previous release branch of Alpine Linux. This
branch has an older version of Chromium and chromedriver [[2]] which
does not have these issues with Selenium, and because chromium is in the
community repository it won't be updated in the 3.20 release branch
unless there is an security fix.

Using the older release branch is preferable to just sticking with the
Docker image that we have now, because then we still get the benefit of
security fixes and patches to packages in the main repository.

[1]: https://github.com/teamcapybara/capybara/issues/2800
[2]: https://pkgs.alpinelinux.org/packages?name=chromium*&branch=v3.20&repo=&arch=x86_64&origin=&flagged=&maintainer=