Fix warnings from Cuprite about options to Node#set
Cuprite is currently warning about ignored `fill_options` to `fill_in`:

> Options passed to Node#set but Cuprite doesn't currently support any - ignoring

In commit 9786371 we added `fill_options` to tell Chrome to clear the
field for the email address for completed forms before filling it in, as
autofill was pre-populating it. I'm pretty sure now that Chrome no
longer autofills fields without prompting; I've been unable to recreate
the issue in the commit. So this commit just removes the options, to
stop the warnings from Cuprite.