Make tests compatible with submission emails sent via AWS SES
This commit adds a NotifyException to NotifyHelpers, which is then used
in the tests to attempt submission via SES if emails cannot be found in
Notify. This means it is possible for the end-to-end tests to work with
submission emails sent via either Notify or AWS SES. Notify is
attempted first as this is the existing behaviour.