Remove branch feature flag conditional text
To support the branch feature flag, we made the tests work for when the
feature flag is enabled or disabled.

This was done by replacing text with regexes.

This commit replaces the conditional text with the current text values
now that that feature flag has been removed and the branching feature
released.