Change privacy policy URL in created form tests
The privacy policy URL in form the tests create uses the example URL
shown on the page.

The validation for the privacy policy URL has been changed to stop users
entering the example URL.

This commit changes the privacy policy URL in the tests to a fictional
URL which passes the validation.