Add Cuprite as optional alternative to ChromeDriver
If USE_CUPRITE=1 then use [Cuprite] instead of ChromeDriver.

I've had some success using Cuprite locally, but we had to pull it from
this codebase when we last tried it in PR #120 because it wasn't working
reliably in our automated test and deployment pipelines.

This commit adds it back, but as an optional feature, currently intended
only for local developer environments.

By default `bundle install` will install cuprite, but if for some reason
it is unwanted it can be ignored by telling Bundler to go without the
`cuprite` group:

    bundle config set --local without cuprite

[Cuprite]: https://github.com/rubycdp/cuprite