Add aws-cli to Docker image
We use this Docker image in AWS CodeBuild, and it is useful to have the
AWS CLI present to run commands around the test run without having to
install it first.

This commit adds aws-cli to the installed packages for this image; this
adds roughly 200 MiB of disk space while the image is in use, but only
20 MiB to the content size, and saves bandwidth/run time whenever the
end to end tests are run.