Fix deleting forms with admin user
When running the tests locally the delete form step would fail due to
an ambiguous match.

This was happening because when running locally with SKIP_AUTH=1 the
tests would use the default user, which is usually a super admin, and
this user sees the 'change group' link for each form on the group forms
page, and that change group link has the name of form in hidden text
which matches the query used previously.

This commit fixes things by telling the resolver to require an exact
match with exact text, so it will not match links that include the name
of the form in the link text, but will still warn if there is more than
one link with the same form name on the page.

Unfortunately it's not clear why exactly the `:one` matcher was needed
originally, but I think this change should preserve the original need.