Add openssl gem to development group
When running against AWS locally there is a certificate error with the
version of the OpenSSL gem bundled with Ruby 3.4.4 and the latest
OpenSSL library (see https://github.com/ruby/openssl/issues/949).
Bumping the version of the OpenSSL gem used fixes thing.

However, building the OpenSSL gem in the container requires extra
dependencies, and the errors don't appear to happen there (probably due
to an older version of the OpenSSL library), so we mark the OpenSSL gem
as a development gem only.