Refactor add question methods for reusability
Add a method `add_a_question` that gets us to the "What kind of answer
do you need?" page if we're not there already, regardless of whether the
form has no questions yet, or we're on the add and edit questions page,
or on the page to edit a question that has just been saved. This lets us
make the methods to add questions look more uniform.

Also rename the methods to be `add_a_{type}_question`, to match the
language in the service.