Add configuration for settings with config gem
Add the files required to use the config gem (mostly copied from our
Rails apps), plus the code to load the gem and settings into the tests,
along with scaffolding/default settings.

The default environment is called "development", to align with the
development environment/settings in our Rails apps. This refers to a
developer environment with apps running locally, rather than the "dev"
environment in AWS.