Fix Dockerfile test script
Unset the `-u` shell attribute which raises errors for unbound
variables; otherwise the script does not work without arguments given and all
environment variables set, even though it is written to be able to work
without that.

Also make sure the `load_env_vars` script is sourced with an absolute
path, so that the dockerfile_test script does not need to be run in the
`bin` directory.