Add forms environment envvar
Add an environment variable with the name of the AWS environment the end
to end tests are being run against. The name of the environment variable
is based on that set for our apps in the forms-deploy repo.

Nothing in our end to end tests themselves use this currently, but it
will become more useful later on.