Use a more realistic file for testing file uploads
Adds a medium sized image to be uploaded in answer to the file upload
question in the end to end test form. This exercises our system more
completely than the previous test file, which was very small and not
representative of the sorts of uploads we expect.

The image was taken by the committer and is released under the usual
licence.