Move steps for building a new form into end to end spec file
The `build_a_new_form` method is only used in one test in the end to end
spec file, so it doesn't really need to be in the feature helper file.
Plus, expanding the steps instead of having just one method makes it a
bit easier to see what the tests do and where to add new steps.